+++ packages/coding-agent/test/suite/regressions/5943-session-start-notify.test.ts
@@ -15,6 +15,7 @@ function createUiContext(
 		confirm: async () => false,
 		input: async () => undefined,
 		notify: onNotify,
+		transientStatus: () => {},
 		onTerminalInput: () => () => {},
 		setStatus: () => {},
 		setWorkingMessage: () => {},